arm64: defconfig: Disable qbt1000 for SOMC kumano
SOMC kumano's volume down rocker is mapped to qbt1000 within qcom's DTS
instead of a normal volume rocker. By default, this results in a button
that simply triggers a GPIO:

  qbt1000 : gpio_report_event: gpio 1273: report state 0
  qbt1000 : gpio_report_event: gpio 1273: report state 1
  qbt1000 : gpio_report_event: gpio 1273: report state 0
  qbt1000 : gpio_report_event: gpio 1273: report state 1

Disable CONFIG_MSM_QBT1000 in kumano's defconfig fragment to fix that,
so that upstream can mess with the DTS without bothering the platform.

Fixes: 4531a60fae9f08d328bd ("arm64: defconfig: Move kumano to config fragments")
Fixes: 7a67a8d2baa4903c1973 ("Update bahamut defconfig")
Signed-off-by: voidanix <voidanix@keyedlimepie.org>